made a validateParent and child file
