+++ middleware/validate.js
@@ -1,4 +1,3 @@
-const { Parent } = require('../models/Model')
 /* check for all credentials for register
  * parent_name
  * username
@@ -49,32 +48,7 @@ function validateLogin() {
   }
 }
 
-/* check for id
- * make sure the id exists
- */
-function validateParentId() {
-  return async (req, res, next) => {
-    try {
-      const [parent] = await Parent.findBy({ id: req.params.id })
-      if (!parent) {
-        return res.status(404).json({ message: 'Parent not found' })
-      } else {
-        let { password, ...rest } = parent
-        req.parent = { ...rest }
-        next()
-      }
-    } catch (error) {
-      next(error)
-    }
-  }
-}
-
-/*
- * Check for all fields on update
- */
-
 module.exports = {
   validateRegister,
-  validateLogin,
-  validateParentId
+  validateLogin
 }
